refactor(vinted): decompose VintedCheckItem god-component (v1.15.2)
God-Component decomposition, step 2/5. VintedCheckItem shrinks from 528
to 75 lines; the parent is now composition + data-source selection +
cross-domain handlers only.

Extracted into src/components/stats/vinted/:
- VintedScanControls — header, resume checkbox, debug toggle, ritual buttons
- VintedScanProgress — active-scan progress bar
- VintedDebugLog — collapsible per-attempt diagnostics panel
- VintedResolveStatus — seller-resolution progress + stored-source banners
- VintedBundleList — seller bundles; owns its own bundleSort state
- VintedBookResultList — per-book result grid (+ internal OfferRow)

Behavior-preserving; 253 tests still pass, lint + build clean.

Refs #91
